+++ addOns/authhelper/CHANGELOG.md
@@ -30,6 +30,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Clear fields before sending keys for Browser Based Authentication, including when using steps.
 - Do not add an empty line to the start of the Other Info of Session Management Response Identified scan rule's alerts.
 - Update the Client Script Based Authentication help page with the new Automation Framework `scriptInline` field.
+- The Authentication Request Detection and Session Management Detection scan rules now skip resources (images, css, js, etc) which are unlikely to be relevant.
+- The Verification Detection scan rule now skips messages that seem related to login/logout/registration functionality.
 
 ### Fixed
 - Correct descriptions of the Zest script steps in the Authentication Report.